NODE-412: Code review: Move synchronization related stuff to 'synchronization' package
