+++ comm/src/test/scala/io/casperlabs/comm/gossiping/synchronization/StashingSynchronizerSpec.scala
@@ -5,7 +5,8 @@ import cats.syntax.either._
 import com.google.protobuf.ByteString
 import io.casperlabs.casper.consensus.BlockSummary
 import io.casperlabs.comm.discovery.Node
-import io.casperlabs.comm.gossiping.Synchronizer.SyncError
+import io.casperlabs.comm.gossiping.synchronization.Synchronizer.SyncError
+import io.casperlabs.comm.gossiping.synchronization.Synchronizer
 import monix.eval.Task
 import monix.execution.atomic.{Atomic, AtomicInt}
 import monix.execution.schedulers.CanBlock.permit
